[vnet] install and run windows service
This commit adds a Windows service for VNet. It adds support for
automatically installing and running the service when the user runs
`tsh vnet`, and adds a command to manually uninstall/delete the service.

The service creates the TUN interface and establishes an IPC connection
with the user process over a named pipe, but for now does not actually
handle any networking, the rest will come in later PRs.

If you want to test this out on a Windows machine/VM, you should be able
to run `tsh vnet` and see that:
1. A service TeleportVNet is installed and runs
   `sc.exe query state=all | grep -A3 Teleport`
2. The service writes logs to `logs.txt` in the directory where `tsh` is
   installed (this is temporary until I find a better place for logs).
3. A TUN interface is created `netsh interface show interface`
4. The service stops and the interface is cleaned up when the user
   process exits.

Unfortunately this PR does not include any unit tests. Most of the
functionality here needs to be able to escalate to administrator with a
UAC prompt and needs to run on Windows, this is exactly the kind of unit
test that is very hard to write and would never actually be able to run
in CI.
But, any part of this that's broken would immediately break VNet on
Windows, and this should be caught in any test plan.